nigative total

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        long ballance = 2_000_000_000;
-        long transfer = 500_000_000;
-        long total = ballance * transfer;
+        int ballance = 2_000_000_000;
+        int transfer = 500_000_000;
+        int total = ballance * transfer;
         System.out.println(total);
 
     }
